+++ src/frontend/src/binder/relation/join.rs
@@ -1,0 +1,87 @@
+use risingwave_common::error::{ErrorCode, Result};
+use risingwave_common::types::DataType;
+use risingwave_pb::plan::JoinType;
+use risingwave_sqlparser::ast::{JoinConstraint, JoinOperator, TableWithJoins};
+
+use crate::binder::{Binder, Relation};
+use crate::expr::{Expr as _, ExprImpl};
+
+#[derive(Debug)]
+pub struct BoundJoin {
+    pub join_type: JoinType,
+    pub left: Relation,
+    pub right: Relation,
+    pub cond: ExprImpl,
+}
+
+impl Binder {
+    pub(crate) fn bind_vec_table_with_joins(
+        &mut self,
+        from: Vec<TableWithJoins>,
+    ) -> Result<Option<Relation>> {
+        let mut from_iter = from.into_iter();
+        let first = match from_iter.next() {
+            Some(t) => t,
+            None => return Ok(None),
+        };
+        let mut root = self.bind_table_with_joins(first)?;
+        for t in from_iter {
+            let right = self.bind_table_with_joins(t)?;
+            root = Relation::Join(Box::new(BoundJoin {
+                join_type: JoinType::Inner,
+                left: root,
+                right,
+                cond: ExprImpl::literal_bool(true),
+            }));
+        }
+        Ok(Some(root))
+    }
+
+    fn bind_table_with_joins(&mut self, table: TableWithJoins) -> Result<Relation> {
+        let mut root = self.bind_table_factor(table.relation)?;
+        for join in table.joins {
+            let right = self.bind_table_factor(join.relation)?;
+            let (constraint, join_type) = match join.join_operator {
+                JoinOperator::Inner(constraint) => (constraint, JoinType::Inner),
+                JoinOperator::LeftOuter(constraint) => (constraint, JoinType::LeftOuter),
+                JoinOperator::RightOuter(constraint) => (constraint, JoinType::RightOuter),
+                JoinOperator::FullOuter(constraint) => (constraint, JoinType::FullOuter),
+                // Cross join equals to inner join with with no constraint.
+                JoinOperator::CrossJoin => (JoinConstraint::None, JoinType::Inner),
+            };
+            let cond = self.bind_join_constraint(constraint)?;
+            let join = BoundJoin {
+                join_type,
+                left: root,
+                right,
+                cond,
+            };
+            root = Relation::Join(Box::new(join));
+        }
+
+        Ok(root)
+    }
+
+    fn bind_join_constraint(&mut self, constraint: JoinConstraint) -> Result<ExprImpl> {
+        Ok(match constraint {
+            JoinConstraint::None => ExprImpl::literal_bool(true),
+            JoinConstraint::Natural => {
+                return Err(ErrorCode::NotImplemented("Natural join".into(), 1633.into()).into())
+            }
+            JoinConstraint::On(expr) => {
+                let bound_expr = self.bind_expr(expr)?;
+                if bound_expr.return_type() != DataType::Boolean {
+                    return Err(ErrorCode::InternalError(format!(
+                        "argument of ON must be boolean, not type {:?}",
+                        bound_expr.return_type()
+                    ))
+                    .into());
+                }
+                bound_expr
+            }
+            JoinConstraint::Using(_columns) => {
+                return Err(ErrorCode::NotImplemented("USING".into(), 1636.into()).into())
+            }
+        })
+    }
+}
